Fix frontend tests and consolidate Bitbucket test files
