--- tests/unit/resolver/bitbucket/test_bitbucket_send_pull_request.py
@@ -1,90 +1,0 @@
-from unittest.mock import MagicMock, patch
-
-from openhands.integrations.service_types import ProviderType
-from openhands.resolver.interfaces.bitbucket import BitbucketIssueHandler
-from openhands.resolver.interfaces.issue import Issue
-from openhands.resolver.interfaces.issue_definitions import ServiceContextIssue
-from openhands.resolver.send_pull_request import send_pull_request
-
-
-@patch('openhands.resolver.send_pull_request.ServiceContextIssue')
-@patch('openhands.resolver.send_pull_request.BitbucketIssueHandler')
-@patch('subprocess.run')
-def test_send_pull_request_bitbucket(
-    mock_run, mock_bitbucket_handler, mock_service_context
-):
-    # Mock subprocess.run to avoid actual git operations
-    mock_run.return_value = MagicMock(returncode=0)
-
-    # Mock the BitbucketIssueHandler instance
-    mock_instance = MagicMock(spec=BitbucketIssueHandler)
-    mock_bitbucket_handler.return_value = mock_instance
-
-    # Mock the ServiceContextIssue instance
-    mock_service = MagicMock(spec=ServiceContextIssue)
-    mock_service.get_branch_name.return_value = 'openhands-fix-123'
-    mock_service.branch_exists.return_value = True
-    mock_service.get_default_branch_name.return_value = 'main'
-    mock_service.get_clone_url.return_value = (
-        'https://bitbucket.org/test-workspace/test-repo.git'
-    )
-    mock_service.create_pull_request.return_value = {
-        'html_url': 'https://bitbucket.org/test-workspace/test-repo/pull-requests/123'
-    }
-    # Add _strategy attribute to mock
-    mock_strategy = MagicMock()
-    mock_service._strategy = mock_strategy
-    mock_service_context.return_value = mock_service
-
-    # Create a mock Issue
-    mock_issue = Issue(
-        number=123,
-        title='Test Issue',
-        owner='test-workspace',
-        repo='test-repo',
-        body='Test body',
-        created_at='2023-01-01T00:00:00Z',
-        updated_at='2023-01-01T00:00:00Z',
-        closed_at=None,
-        head_branch='feature-branch',
-        thread_ids=None,
-    )
-
-    # Call send_pull_request
-    result = send_pull_request(
-        issue=mock_issue,
-        token='test-token',
-        username=None,
-        platform=ProviderType.BITBUCKET,
-        patch_dir='/tmp',  # Use /tmp instead of /tmp/repo to avoid directory not found error
-        pr_type='ready',
-        pr_title='Test PR',
-        target_branch='main',
-    )
-
-    # Verify the result
-    assert result == 'https://bitbucket.org/test-workspace/test-repo/pull-requests/123'
-
-    # Verify the handler was created correctly
-    mock_bitbucket_handler.assert_called_once_with(
-        'test-workspace',
-        'test-repo',
-        'test-token',
-        None,
-        'bitbucket.org',
-    )
-
-    # Verify ServiceContextIssue was created correctly
-    mock_service_context.assert_called_once()
-
-    # Verify create_pull_request was called with the correct data
-    expected_body = 'This pull request fixes #123.\n\nAutomatic fix generated by [OpenHands](https://github.com/All-Hands-AI/OpenHands/) 🙌'
-    mock_service.create_pull_request.assert_called_once_with(
-        {
-            'title': 'Test PR',
-            'description': expected_body,
-            'source_branch': 'openhands-fix-123',
-            'target_branch': 'main',
-            'draft': False,
-        }
-    )
